+++ pkg/client/chunk/observed_test.go
@@ -2,7 +2,7 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package utils
+package chunk
 
 import (
 	"golang.org/x/exp/slices"
@@ -20,7 +20,7 @@ import (
 	"github.com/elastic/elastic-agent-client/v7/pkg/proto"
 )
 
-func TestChunkedObserved(t *testing.T) {
+func TestObserved(t *testing.T) {
 	timestamp := time.Now()
 
 	scenarios := []struct {
@@ -88,8 +88,8 @@ func TestChunkedObserved(t *testing.T) {
 			},
 		},
 		{
-			Name:    "chunk",
-			MaxSize: 100,
+			Name:    "chunk checkin message",
+			MaxSize: 120,
 			Original: &proto.CheckinObserved{
 				Token:        "token",
 				FeaturesIdx:  2,
@@ -112,6 +112,16 @@ func TestChunkedObserved(t *testing.T) {
 						State:          proto.State_HEALTHY,
 						Message:        "Healthy",
 					},
+					{
+						Id:             "id-three",
+						Type:           proto.UnitType_INPUT,
+						ConfigStateIdx: 1,
+						State:          proto.State_HEALTHY,
+						Message:        "Healthy",
+						Payload: mustStruct(map[string]interface{}{
+							"large": "larger than id-two",
+						}),
+					},
 				},
 			},
 			Expected: []*proto.CheckinObserved{
@@ -126,6 +136,16 @@ func TestChunkedObserved(t *testing.T) {
 							ConfigStateIdx: 1,
 							State:          proto.State_HEALTHY,
 							Message:        "Healthy",
+						},
+						{
+							Id:             "id-three",
+							Type:           proto.UnitType_INPUT,
+							ConfigStateIdx: 1,
+							State:          proto.State_HEALTHY,
+							Message:        "Healthy",
+							Payload: mustStruct(map[string]interface{}{
+								"large": "larger than id-two",
+							}),
 						},
 					},
 					UnitsTimestamp: timestamppb.New(timestamp),
@@ -211,7 +231,7 @@ func TestChunkedObserved(t *testing.T) {
 
 	for _, scenario := range scenarios {
 		t.Run(scenario.Name, func(t *testing.T) {
-			observed, err := ChunkedObserved(scenario.Original, scenario.MaxSize, WithTimestamp(timestamp))
+			observed, err := Observed(scenario.Original, scenario.MaxSize, WithTimestamp(timestamp))
 			if scenario.Error != "" {
 				require.Error(t, err)
 				assert.True(t, strings.Contains(err.Error(), scenario.Error))
@@ -221,7 +241,7 @@ func TestChunkedObserved(t *testing.T) {
 				require.Empty(t, diff)
 
 				// re-assemble and it should now match the original
-				assembled, err := RecvChunkedObserved(&fakeCheckinObservedReceiver{msgs: observed})
+				assembled, err := RecvObserved(&fakeCheckinObservedReceiver{msgs: observed})
 				require.NoError(t, err)
 
 				// to compare we need to remove the units timestamp and ensure the units are in the same order
@@ -237,218 +257,88 @@ func TestChunkedObserved(t *testing.T) {
 	}
 }
 
-func TestChunkedExpected(t *testing.T) {
-	timestamp := time.Now()
-
-	scenarios := []struct {
-		Name     string
-		MaxSize  int
-		Original *proto.CheckinExpected
-		Expected []*proto.CheckinExpected
-		Error    string
-	}{
-		{
-			Name:    "unit too large to fit",
-			MaxSize: 30,
-			Error:   "unable to chunk proto.CheckinExpected the unit id-one is larger than max",
-			Original: &proto.CheckinExpected{
-				Units: []*proto.UnitExpected{
-					{
-						Id:             "id-one",
-						Type:           proto.UnitType_OUTPUT,
-						ConfigStateIdx: 1,
-						State:          proto.State_HEALTHY,
-						LogLevel:       proto.UnitLogLevel_INFO,
-						Config: &proto.UnitExpectedConfig{
-							Id:   "testing",
-							Type: "testing",
-							Name: "testing",
-						},
-					},
-					{
-						Id:             "id-two",
-						Type:           proto.UnitType_INPUT,
-						ConfigStateIdx: 1,
-						State:          proto.State_HEALTHY,
-					},
-				},
+func TestRecvObserved_Timestamp_Restart(t *testing.T) {
+	firstTimestamp := time.Now()
+	first := &proto.CheckinObserved{
+		Token:        "token",
+		FeaturesIdx:  2,
+		ComponentIdx: 3,
+		Units: []*proto.UnitObserved{
+			{
+				Id:             "first-one",
+				Type:           proto.UnitType_OUTPUT,
+				ConfigStateIdx: 1,
+				State:          proto.State_HEALTHY,
+				Message:        "Healthy",
+				Payload: mustStruct(map[string]interface{}{
+					"large": "this structure places this unit over the maximum size",
+				}),
 			},
-		},
-		{
-			Name:    "first chunk too large",
-			MaxSize: 50,
-			Error:   "unable to chunk proto.CheckinExpected the first chunk with",
-			Original: &proto.CheckinExpected{
-				Units: []*proto.UnitExpected{
-					{
-						Id:             "id-one",
-						Type:           proto.UnitType_OUTPUT,
-						ConfigStateIdx: 1,
-						State:          proto.State_HEALTHY,
-						LogLevel:       proto.UnitLogLevel_INFO,
-						Config: &proto.UnitExpectedConfig{
-							Id:   "testing1",
-							Type: "testing",
-							Name: "testing1",
-						},
-					},
-					{
-						Id:             "id-two",
-						Type:           proto.UnitType_INPUT,
-						ConfigStateIdx: 1,
-						State:          proto.State_HEALTHY,
-						LogLevel:       proto.UnitLogLevel_INFO,
-						Config: &proto.UnitExpectedConfig{
-							Id:   "testing2",
-							Type: "testing",
-							Name: "testing2",
-						},
-					},
-				},
-			},
-		},
-		{
-			Name:    "chunk",
-			MaxSize: 50,
-			Original: &proto.CheckinExpected{
-				FeaturesIdx:  2,
-				ComponentIdx: 3,
-				Units: []*proto.UnitExpected{
-					{
-						Id:             "id-one",
-						Type:           proto.UnitType_OUTPUT,
-						ConfigStateIdx: 1,
-						State:          proto.State_HEALTHY,
-						Config: &proto.UnitExpectedConfig{
-							Id:   "testing",
-							Type: "testing",
-							Name: "testing",
-						},
-					},
-					{
-						Id:             "id-two",
-						Type:           proto.UnitType_INPUT,
-						ConfigStateIdx: 1,
-						State:          proto.State_HEALTHY,
-					},
-				},
-			},
-			Expected: []*proto.CheckinExpected{
-				{
-					FeaturesIdx:  2,
-					ComponentIdx: 3,
-					Units: []*proto.UnitExpected{
-						{
-							Id:             "id-two",
-							Type:           proto.UnitType_INPUT,
-							ConfigStateIdx: 1,
-							State:          proto.State_HEALTHY,
-						},
-					},
-					UnitsTimestamp: timestamppb.New(timestamp),
-				},
-				{
-					Units: []*proto.UnitExpected{
-						{
-							Id:             "id-one",
-							Type:           proto.UnitType_OUTPUT,
-							ConfigStateIdx: 1,
-							State:          proto.State_HEALTHY,
-							Config: &proto.UnitExpectedConfig{
-								Id:   "testing",
-								Type: "testing",
-								Name: "testing",
-							},
-						},
-					},
-					UnitsTimestamp: timestamppb.New(timestamp),
-				},
-				{
-					Units:          []*proto.UnitExpected{},
-					UnitsTimestamp: timestamppb.New(timestamp),
-				},
-			},
-		},
-		{
-			Name:    "fits in single message",
-			MaxSize: 200,
-			Original: &proto.CheckinExpected{
-				FeaturesIdx:  2,
-				ComponentIdx: 3,
-				Units: []*proto.UnitExpected{
-					{
-						Id:             "id-one",
-						Type:           proto.UnitType_OUTPUT,
-						ConfigStateIdx: 1,
-						State:          proto.State_HEALTHY,
-						Config: &proto.UnitExpectedConfig{
-							Id:   "testing",
-							Type: "testing",
-							Name: "testing",
-						},
-					},
-					{
-						Id:             "id-two",
-						Type:           proto.UnitType_INPUT,
-						ConfigStateIdx: 1,
-						State:          proto.State_HEALTHY,
-					},
-				},
-			},
-			Expected: []*proto.CheckinExpected{
-				{
-					FeaturesIdx:  2,
-					ComponentIdx: 3,
-					Units: []*proto.UnitExpected{
-						{
-							Id:             "id-one",
-							Type:           proto.UnitType_OUTPUT,
-							ConfigStateIdx: 1,
-							State:          proto.State_HEALTHY,
-							Config: &proto.UnitExpectedConfig{
-								Id:   "testing",
-								Type: "testing",
-								Name: "testing",
-							},
-						},
-						{
-							Id:             "id-two",
-							Type:           proto.UnitType_INPUT,
-							ConfigStateIdx: 1,
-							State:          proto.State_HEALTHY,
-						},
-					},
-				},
+			{
+				Id:             "first-two",
+				Type:           proto.UnitType_INPUT,
+				ConfigStateIdx: 1,
+				State:          proto.State_HEALTHY,
+				Message:        "Healthy",
 			},
 		},
 	}
+	firstMsgs, err := Observed(first, 100, WithTimestamp(firstTimestamp))
+	require.NoError(t, err)
 
-	for _, scenario := range scenarios {
-		t.Run(scenario.Name, func(t *testing.T) {
-			observed, err := ChunkedExpected(scenario.Original, scenario.MaxSize, WithTimestamp(timestamp))
-			if scenario.Error != "" {
-				require.Error(t, err)
-				assert.True(t, strings.Contains(err.Error(), scenario.Error))
-			} else {
-				require.NoError(t, err)
-				diff := cmp.Diff(scenario.Expected, observed, protocmp.Transform())
-				require.Empty(t, diff)
-
-				// re-assemble and it should now match the original
-				assembled, err := RecvChunkedExpected(&fakeCheckinExpectedReceiver{msgs: observed})
-				require.NoError(t, err)
-
-				// to compare we need to remove the units timestamp and ensure the units are in the same order
-				// completely acceptable that they get re-ordered in the chunking process
-				assembled.UnitsTimestamp = nil
-				slices.SortStableFunc(assembled.Units, sortExpectedUnits)
-				slices.SortStableFunc(scenario.Original.Units, sortExpectedUnits)
-
-				diff = cmp.Diff(scenario.Original, assembled, protocmp.Transform())
-				assert.Empty(t, diff)
-			}
-		})
+	secondTimestamp := time.Now()
+	second := &proto.CheckinObserved{
+		Token:        "token",
+		FeaturesIdx:  2,
+		ComponentIdx: 3,
+		Units: []*proto.UnitObserved{
+			{
+				Id:             "second-one",
+				Type:           proto.UnitType_OUTPUT,
+				ConfigStateIdx: 1,
+				State:          proto.State_HEALTHY,
+				Message:        "Healthy",
+				Payload: mustStruct(map[string]interface{}{
+					"large": "this structure places this unit over the maximum size",
+				}),
+			},
+			{
+				Id:             "second-two",
+				Type:           proto.UnitType_INPUT,
+				ConfigStateIdx: 1,
+				State:          proto.State_HEALTHY,
+				Message:        "Healthy",
+			},
+		},
 	}
+	secondsMsgs, err := Observed(second, 100, WithTimestamp(secondTimestamp))
+	require.NoError(t, err)
+
+	// ensure chunking results in exact length as the order in the test relies on it
+	require.Len(t, firstMsgs, 3)
+	require.Len(t, secondsMsgs, 3)
+
+	// re-order the messages
+	reorderedMsgs := make([]*proto.CheckinObserved, 6)
+	reorderedMsgs[0] = firstMsgs[0]
+	reorderedMsgs[1] = secondsMsgs[0] // becomes new set
+	reorderedMsgs[2] = firstMsgs[1]   // ignored
+	reorderedMsgs[3] = firstMsgs[2]   // ignored
+	reorderedMsgs[4] = secondsMsgs[1]
+	reorderedMsgs[5] = secondsMsgs[2]
+
+	// re-assemble and it should now match the second
+	assembled, err := RecvObserved(&fakeCheckinObservedReceiver{msgs: reorderedMsgs})
+	require.NoError(t, err)
+
+	// to compare we need to remove the units timestamp and ensure the units are in the same order
+	// completely acceptable that they get re-ordered in the chunking process
+	assembled.UnitsTimestamp = nil
+	slices.SortStableFunc(assembled.Units, sortObservedUnits)
+	slices.SortStableFunc(second.Units, sortObservedUnits)
+
+	diff := cmp.Diff(second, assembled, protocmp.Transform())
+	assert.Empty(t, diff)
 }
 
 func mustStruct(v map[string]interface{}) *structpb.Struct {
@@ -463,26 +353,12 @@ func sortObservedUnits(a *proto.UnitObserved, b *proto.UnitObserved) int {
 	return strings.Compare(a.Id, b.Id)
 }
 
-func sortExpectedUnits(a *proto.UnitExpected, b *proto.UnitExpected) int {
-	return strings.Compare(a.Id, b.Id)
-}
-
 type fakeCheckinObservedReceiver struct {
 	msgs []*proto.CheckinObserved
 }
 
 func (f *fakeCheckinObservedReceiver) Recv() (*proto.CheckinObserved, error) {
 	var msg *proto.CheckinObserved
-	msg, f.msgs = f.msgs[0], f.msgs[1:]
-	return msg, nil
-}
-
-type fakeCheckinExpectedReceiver struct {
-	msgs []*proto.CheckinExpected
-}
-
-func (f *fakeCheckinExpectedReceiver) Recv() (*proto.CheckinExpected, error) {
-	var msg *proto.CheckinExpected
 	msg, f.msgs = f.msgs[0], f.msgs[1:]
 	return msg, nil
 }
